Calculates quantiles of a track expression

Calculates quantiles of a track expression for the given percentiles.

emr_quantiles(
  expr,
  percentiles = 0.5,
  stime = NULL,
  etime = NULL,
  iterator = NULL,
  keepref = F,
  filter = NULL
)

Arguments

expr
track expression

percentiles
an array of percentiles of quantiles in [0, 1] range

stime
start time scope

etime
end time scope

iterator
track expression iterator. If 'NULL' iterator is determined implicitly based on track expression.

keepref
If 'TRUE' references are preserved in the iterator.

filter
Iterator filter.

Value

An array that represent quantiles.

Details

This function calculates quantiles for the given percentiles.

If data size exceeds the limit (see: 'getOption(emr_max.data.size)'), the data is randomly sampled to fit the limit. A warning message is generated then.

See also

Examples

emr_db.init_examples()
emr_quantiles("sparse_track", c(0.1, 0.6, 0.8))

#>   0.1   0.6   0.8 
#>  13.0  90.8 168.0