Finds Id-Time points that match track expression

Finds all patient-time pairs where track expression is 'TRUE'.

emr_screen(
  expr,
  sort = F,
  stime = NULL,
  etime = NULL,
  iterator = NULL,
  keepref = F,
  filter = NULL
)

Arguments

expr
logical track expression

sort
if 'TRUE' result is sorted by id, time and reference

stime
start time scope

etime
end time scope

iterator
track expression iterator. If 'NULL' iterator is determined implicitly based on track expression.

keepref
If 'TRUE' references are preserved in the iterator.

filter
Iterator filter.

Value

A set of Id-Time points that match track expression.

Details

This function finds all Id-Time points where track expression's value is 'TRUE'.

If 'sort' is 'TRUE' the returned value is sorted by id, time and reference, otherwise the order is not guaranteed especially for longer runs, when multitasking might be launched. Sorting requires additional time, so it is switched off by default.

See also

Examples

emr_db.init_examples()
emr_screen("sparse_track == 13 | dense_track < 80",
    iterator = "sparse_track", keepref = T
)

#>    id time ref
#> 1   5    1   3
#> 2  10    1   3
#> 3  25    1   0
#> 4  25    2   0
#> 5  25    2   2
#> 6  25    2   6
#> 7  25    2   8
#> 8  25    3   4
#> 9  25    6   0
#> 10 25    6   2
#> 11 40    1   3