Adds new records to a track

Adds new records to a track from a TAB-delimited file or a data frame.

emr_track.addto(track, src)

Arguments

track
track name

src
file name or data-frame containing the track records

Details

This function adds new records to a track. The records are contained either in a file or a data frame.

If 'src' is a file name, the latter must be constituted of four columns separated by spaces or 'TAB' characters: ID, time, reference and value. The file might contain lines of comments which should start with a '#' character.

Alternatively 'src' can be a data frame consisting of the columns named "id", "time", "ref" and "value". Note: "ref" column in the data frame is optional.

Value

None.

See also