Generate a default name for a virtual track

Given virtual track parameters, generate a name with the following format: "vt_src.func_func.params_params.krkeepref.ts_time.shift.id_id.map.filter_filter" Where for 'params', 'time.shift', and 'id.map', the values are separated by an underscore.

emr_vtrack.name(
  src,
  func = NULL,
  params = NULL,
  keepref = FALSE,
  time.shift = NULL,
  id.map = NULL,
  filter = NULL
)

Arguments

src

a character vector specifying the source dataset(s) or filter(s) that the virtual track is based on

func

a character vector specifying the function(s) applied to the source data to generate the virtual track

params

a named list specifying the parameters used by the function(s) to generate the virtual track

keepref

a logical value indicating whether the virtual track should keep the reference column(s) of the source data

time.shift

a numeric vector specifying the time shift(s) applied to the virtual track

id.map

a named list specifying the mapping of the IDs between the source data and the virtual track

filter

a character vector specifying the filter(s) applied to the virtual track. Note that the filter name cannot contain the character '.'

Value

a default name for the virtual track

Details

If func, params, time.shift, id.map, or filter are NULL - their section would not appear in the generated name.

See also

Examples

emr_db.init_examples()
#> NULL
emr_vtrack.name("dense_track",
    time.shift = 1,
    func = "max"
)
#> [1] "vt_dense_track.func_max.params_.krF.ts_1"