Calculates correlation statistics for pairs of track expressions

Calculates correlation statistics for pairs of track expressions.

emr_cor(
  ...,
  cor.exprs = NULL,
  include.lowest = FALSE,
  right = TRUE,
  stime = NULL,
  etime = NULL,
  iterator = NULL,
  keepref = F,
  filter = NULL,
  dataframe = FALSE,
  names = NULL
)

Arguments

cor.exprs
vector of track expressions for which correlation statistics is calculated.

include.lowest
if 'TRUE', the lowest (or highest, for ‘right = FALSE’) value of the range determined by breaks is included.

right
if 'TRUE' the intervals are closed on the right (and open on the left), otherwise vice versa.

stime
start time scope.

etime
end time scope.

iterator
track expression iterator. If 'NULL' iterator is determined implicitly based on track expressions.

keepref
If 'TRUE' references are preserved in the iterator.

filter
Iterator filter.

dataframe
return a data frame instead of an N-dimensional vector.

names
names for track expressions in the returned dataframe (only relevant when dataframe == TRUE)

expr
track expression.

breaks
breaks that determine the bin or 'NULL'.

Value

A list of 5 elements each containing a N-dimensional vector (N is the number of 'expr'-'breaks' pairs). The member of each vector is a specific statistics matrix. If dataframe == TRUE - a data frame with a column for each track expression, additional columns i,j with pairs of cor_exprs and another 5 columns: 'n', 'e', 'var', 'cov', 'cor', see description.

Details

This function works in a similar manner to 'emr_dist'. However instead of returning a single counter for each bin 'emr_cor' returns 5 matrices of 'length(cor.exprs) X length(cor.exprs)' size. Each matrix represents the correlation statistics for each pair of track expressions from 'cor.exprs'. Given a 'bin' and a pair of track expressions 'cor.exprs[i]' and 'cor.exprs[j]' the corresponding matrix contains the following information:

$n[bin,i,j] - number of times when both 'cor.exprs[i]' and 'cor.exprs[j]' exist $e[bin,i,j] - expectation (average) of values from 'cor.exprs[i]' when 'cor.exprs[j]' exists $var[bin,i,j] - variance of values from 'cor.exprs[i]' when 'cor.exprs[j]' exists $cov[bin,i,j] - covariance of 'cor.exprs[i]' and 'cor.exprs[j]' $cor[bin,i,j] - correlation of 'cor.exprs[i]' and 'cor.exprs[j]'

Similarly to 'emr_dist' 'emr_cor' can do multi-dimensional binning. Given N dimensional binning the individual data in the matrices can be accessed as: $cor[bin1, ..., binN, i, j].

If dataframe = TRUE the return value is a data frame with a column for each track expression, additional columns i,j with pairs of cor_exprs and another 5 columns: 'n', 'e', 'var', 'cov', 'cor' with the same values as the matrices described above.

See also

Examples

emr_db.init_examples()
emr_cor("categorical_track", c(0, 2, 5),
    cor.exprs = c("sparse_track", "1/dense_track"),
    include.lowest = T, iterator = "categorical_track",
    keepref = T
)

#> $n
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            3             3
#> (2,5]            2             2
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            3             4
#> (2,5]            2             2
#> 
#> 
#> $e
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]           48    0.02608341
#> (2,5]          130    0.02136752
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]           48    0.03879333
#> (2,5]          130    0.02136752
#> 
#> 
#> $var
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667  0.0001876685
#> (2,5]   10816.0000  0.0002922054
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667  0.0006253773
#> (2,5]   10816.0000  0.0002922054
#> 
#> 
#> $cov
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667    -0.2520039
#> (2,5]   10816.0000    -1.7777778
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]   -0.2520039  0.0006253773
#> (2,5]   -1.7777778  0.0002922054
#> 
#> 
#> $cor
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            1    -0.9995979
#> (2,5]            1    -1.0000000
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]   -0.9995979             1
#> (2,5]   -1.0000000             1
#> 
#> 
#> attr(,"breaks")
#> attr(,"breaks")[[1]]
#> [1] 0 2 5
#> 

emr_cor("categorical_track", c(0, 2, 5),
    cor.exprs = c("sparse_track", "1/dense_track"),
    include.lowest = T, iterator = "categorical_track",
    keepref = T,
    dataframe = TRUE
)

#> $n
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            3             3
#> (2,5]            2             2
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            3             4
#> (2,5]            2             2
#> 
#> 
#> $e
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]           48    0.02608341
#> (2,5]          130    0.02136752
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]           48    0.03879333
#> (2,5]          130    0.02136752
#> 
#> 
#> $var
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667  0.0001876685
#> (2,5]   10816.0000  0.0002922054
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667  0.0006253773
#> (2,5]   10816.0000  0.0002922054
#> 
#> 
#> $cov
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]     338.6667    -0.2520039
#> (2,5]   10816.0000    -1.7777778
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]   -0.2520039  0.0006253773
#> (2,5]   -1.7777778  0.0002922054
#> 
#> 
#> $cor
#> , , sparse_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]            1    -0.9995979
#> (2,5]            1    -1.0000000
#> 
#> , , 1/dense_track
#> 
#>       sparse_track 1/dense_track
#> [0,2]   -0.9995979             1
#> (2,5]   -1.0000000             1
#> 
#> 
#> attr(,"breaks")
#> attr(,"breaks")[[1]]
#> [1] 0 2 5
#>